Search Export and Delete selectable fix - Reviewed By Graeme

+++ app/webroot/js/search.js
@@ -15,25 +15,10 @@ $(function() {
                 info.push($(this).attr("data-id"));
             });
             window.location = "/search/export?ids=" + info;
-
-
-            jQuery('.ideablock').each(function() {
-                $(this).unbind("click");
-                $(this).attr("onclick", "Ajax.Idea.showIdea($(this).attr(\"data-id\"));");
-                $(this).selectable();
-            });
-            $("#btn-cancel-csv").hide();
-            $("#export-help").hide();
-            $("#btn-save-csv").hide();
-            $("#btn-export").removeAttr("disabled");
+            cancelcsv();
         });
     });
     $("#btn-cancel-csv").click(cancelcsv);
-
-
-
-
-
 
     $("#btn-delete").click(function(e) {
         $("#btn-delete").attr("disabled", "disabled");
@@ -49,29 +34,27 @@ $(function() {
             jQuery('.ideablock.selected').each(function() {
                 info.push($(this).attr("data-id"));
             });
-                bootbox.confirm({
-                message: "Are you sure you want to delete this idea?", 
-                    buttons: {
-                        confirm: {label: "Delete"},
-                        cancel: {label: "Don't do it!"}
-                }, 
-                    callback:function(result) {  
-                        if(result === true){
-                            Ajax.Idea.deleteIdea(info); 
-                        }
-           }
-}).find("div.modal-content").addClass("confirmWidth"); 
+            bootbox.confirm({
+                message: "Are you sure you want to delete this idea?",
+                buttons: {
+                    confirm: {label: "Delete"},
+                    cancel: {label: "Don't do it!"}
+                },
+                callback:function(result) {
+                    if(result === true){
+                        Ajax.Idea.deleteIdea(info);
+                    }
+                }
+            }).find("div.modal-content").addClass("confirmWidth");
         });
     });
     $("#btn-cancel-delete").click(canceldelete);
-
-
 
     function cancelcsv(e) {
         jQuery('.ideablock').each(function() {
             $(this).unbind("click");
             $(this).attr("onclick", "Ajax.Idea.showIdea($(this).attr(\"data-id\"));");
-            $(this).selectable();
+            $(this).selectable(false);
         });
         $("#btn-cancel-csv").hide();
         $("#export-help").hide();
@@ -80,18 +63,13 @@ $(function() {
     }
 
     function canceldelete(e) {
-            jQuery('.ideablock').each(function() {
-                $(this).unbind("click");
-                $(this).attr("onclick", "Ajax.Idea.showIdea($(this).attr(\"data-id\"));");
-                $(this).selectable();
-            });
-            $("#btn-cancel-delete").hide();
-            $("#btn-delete-confirm").hide();
-            $("#btn-delete").removeAttr("disabled");
+        jQuery('.ideablock').each(function() {
+            $(this).unbind("click");
+            $(this).attr("onclick", "Ajax.Idea.showIdea($(this).attr(\"data-id\"));");
+            $(this).selectable(false);
+        });
+        $("#btn-cancel-delete").hide();
+        $("#btn-delete-confirm").hide();
+        $("#btn-delete").removeAttr("disabled");
     }
-
-
 });
-
-
-
